feat: move and add box-shadow picker component
