+++ studio/src/app/components/editor/styles/element/app-border-radius/app-border-radius.tsx
@@ -5,7 +5,7 @@ import {RangeChangeEventDetail} from '@ionic/core';
 @Component({
   tag: 'app-border-radius',
 })
-export class BorderRadius {
+export class AppBorderRadius {
   @Prop()
   selectedElement: HTMLElement;
